separate Leaderboard from LeaderboardManager
