+++ src/RA_AchievementOverlay.cpp
@@ -9,7 +9,7 @@
 #include "RA_ImageFactory.h"
 #include "RA_PopupWindows.h"
 #include "RA_Core.h"
-#include "RA_Leaderboard.h"
+#include "RA_LeaderboardManager.h"
 #include "RA_GameData.h"
 
 #include <time.h>
